framework/summer-web/src/main/java/com/itranswarp/summer/web/utils/PathUtils.java
package com.itranswarp.summer.web.utils;

import java.util.regex.Pattern;

import jakarta.servlet.ServletException;

public class PathUtils {

    public static Pattern compile(String path) throws ServletException {
        String regPath = path.replaceAll("\\{([a-zA-Z][a-zA-Z0-9]*)\\}", "(?<$1>[^/]*)");
        if (regPath.indexOf('{') >= 0 || regPath.indexOf('}') >= 0) {
            throw new ServletException("Invalid path: " + path);
        }
        return Pattern.compile("^" + regPath + "$");
    }
}
